Update to MAOOAM v1.2:
Lua code:
Reduce size of noise for random IC to avoid blow-up.
Don’t rewrite IC file when resuming a previous run.

Fortran code:
Selection of the integration scheme now done in the Makefile
Update of the Makefile
Correction of a bug in initial conditions (IC(0) should be always equal to 1)